+++ admin/forum.php
@@ -43,14 +43,14 @@ if(!isset($_SESSION['Center_Username']) or $_SESSION['Center_UserGroup'] != 9){
     exit;
 }
 
-if(isset($_POST['del']) && $_POST['del'] != ''){
+if(isset($_POST['del']) && $_POST['del'] != '' && isset($_GET[$_SESSION['Center_Auth']])){
     $_del[] = sprintf("DELETE FROM `forum` WHERE `id` IN (%s)",implode(",",$_POST['del']));
     $_del[] = sprintf("DELETE FROM `forum_reply` WHERE `post_id` IN (%s)",implode(",",$_POST['del']));
     foreach($_del as $val){
 		$SQL->query($val);
 	}
 	header("Location: forum.php?del&fid=".$_GET['fid']);
-}elseif(isset($_GET['delblock']) && abs($_GET['delblock']) != ''){
+}elseif(isset($_GET['delblock']) && abs($_GET['delblock']) != '' && isset($_GET[$_SESSION['Center_Auth']])){
 	$_post=sc_get_result("SELECT `id` FROM `forum` WHERE `block` = '%d'",array(abs($_GET['delblock'])));
 	if($_post['num_rows']>0){
 		do{
@@ -65,11 +65,11 @@ if(isset($_POST['del']) && $_POST['del'] != ''){
 		$SQL->query($val);
 	}
 	header("Location: forum.php?del");
-}elseif(isset($_GET['newblock'])&&sc_namefilter($_POST['blockname'])!=''){
+}elseif(isset($_GET['newblock'])&&sc_namefilter($_POST['blockname'])!='' && isset($_GET[$_SESSION['Center_Auth']])){
 
 	sc_add_forum_block(sc_namefilter($_POST['blockname']));
 	
-}elseif(isset($_GET['edit']) &&abs($_GET['edit'])!='' && isset($_POST['blockname'])&&isset($_POST['position'])){
+}elseif(isset($_GET['edit']) &&abs($_GET['edit'])!='' && isset($_POST['blockname'])&&isset($_POST['position']) && isset($_GET[$_SESSION['Center_Auth']])){
 	$SQL->query("UPDATE `forum_block` SET `blockname` = '%s',`position` = '%d' WHERE `id` = '%d'",array(sc_namefilter($_POST['blockname']),abs(intval($_POST['position'])),abs($_GET['edit'])));
 	$_GET['edit']=false;
 }
@@ -82,14 +82,37 @@ if(isset($_GET['fid'])){
 		header("Location: forum.php");
 	}
 	
+	if(!isset($_GET['sort'])){
+		$_GET['sort']='01';
+	}
+	if(isset($_GET['sort'])){
+		$_GET['sort']=intval($_GET['sort']);
+		if(strlen($_GET['sort'])!=2){
+			$_GET['sort']=str_pad($_GET['sort'],2,0,STR_PAD_LEFT);
+		}
+		$_table=array('mktime','title','author');
+		$_a=str_split($_GET['sort'],1);
+		if(!isset($_table[$_a[0]])){
+			$_a[0]=0;
+		}
+		
+		$_sort='`'.$_table[$_a[0]].'` ';
+		
+		if($_a[1]==1){
+			$_sort.='DESC';
+		}else{
+			$_sort.='ASC';
+		}
+	}
+	
 	$limit_row=$center['forum']['limit'];
 	
 	if(isset($_GET['page'])){
 		$limit_start = abs(intval(($_GET['page']-1)*$limit_row));
-		$_forum = sc_get_result("SELECT * FROM `forum` WHERE `block`='%d' ORDER BY `mktime` DESC LIMIT %d,%d",array($_block['row']['id'],$limit_start,$limit_row));
+		$_forum = sc_get_result("SELECT * FROM `forum` WHERE `block`='%d' ORDER BY %s LIMIT %d,%d",array($_block['row']['id'],$_sort,$limit_start,$limit_row));
 	} else {
 		$limit_start=0;
-		$_forum = sc_get_result("SELECT * FROM `forum` WHERE `block`='%d' ORDER BY `mktime` DESC LIMIT %d,%d",array($_block['row']['id'],$limit_start,$limit_row));
+		$_forum = sc_get_result("SELECT * FROM `forum` WHERE `block`='%d' ORDER BY %s LIMIT %d,%d",array($_block['row']['id'],$_sort,$limit_start,$limit_row));
 	}
 }else{
 	$_forum = sc_get_result("SELECT * FROM `forum_block` ORDER BY `position` ASC");
@@ -118,14 +141,14 @@ $(function(){
 	});
 });
 </script>
-<form action="forum.php?fid=<?php echo abs($_GET['fid']); ?>" method="POST">
+<form action="forum.php?fid=<?php echo abs($_GET['fid']).'&'.$_SESSION['Center_Auth']; ?>" method="POST">
 	<div class="table-responsive">
 		<table class="table table-striped table-hover">
 			<thead>
 				<tr>
 					<th width="30"><input class="btn btn-danger btn-sm" type="submit" value="刪除"></th>
-					<th>文章</th>
-					<th>作者/發表時間</th>
+					<th><a href="?fid=<?php echo abs($_GET['fid']); ?>&sort=1<?php if($_a[0]==1)echo ($_a[1]+1)%2; else echo 0; ?>">標題<?php if($_a[0]==1){ ?><span class="glyphicon glyphicon-menu-<?php if($_a[1]==0){ ?>down<?php }else{ ?>up<?php } ?>"></span><?php } ?></a></th>
+					<th><a href="?fid=<?php echo abs($_GET['fid']); ?>&sort=2<?php if($_a[0]==2)echo ($_a[1]+1)%2; else echo 0; ?>">作者<?php if($_a[0]==2){ ?><span class="glyphicon glyphicon-menu-<?php if($_a[1]==0){ ?>down<?php }else{ ?>up<?php } ?>"></span><?php } ?></a>/<a href="?fid=<?php echo abs($_GET['fid']); ?>&sort=0<?php if($_a[0]==0)echo ($_a[1]+1)%2; else echo 0; ?>">發表時間<?php if($_a[0]==0){ ?><span class="glyphicon glyphicon-menu-<?php if($_a[1]==0){ ?>down<?php }else{ ?>up<?php } ?>"></span><?php } ?></a></th>
 					<th>回覆</th>
 					<th>最後回覆</th>
 					<th></th>
@@ -173,7 +196,7 @@ $(function(){
 </form>
 <?php
 	$_all_forum=sc_get_result("SELECT COUNT(*) FROM `forum` WHERE `block`='%d'",array($_block['row']['id']));
-	echo sc_page_pagination('forum.php',@$_GET['page'],implode('',$_all_forum['row']),$center['forum']['limit'],'&fid='.$_block['row']['id']);
+	echo sc_page_pagination('forum.php',@$_GET['page'],implode('',$_all_forum['row']),$center['forum']['limit'],'&fid='.$_block['row']['id'].'&sort='.$_GET['sort']);
 }
 ?>
 <?php
@@ -184,7 +207,7 @@ $(function(){
 	}
 ?>
 <h2 class="page-header">區塊編輯</h2>
-<form class="form-horizontal form-sm" action="forum.php?edit=<?php echo $_block['row']['id']; ?>" method="POST">
+<form class="form-horizontal form-sm" action="forum.php?edit=<?php echo $_block['row']['id'].'&'.$_SESSION['Center_Auth']; ?>" method="POST">
 	<div class="form-group">
 			<label class="col-sm-3 control-label" for="blockname">區塊名稱：</label>
 			<div class="col-sm-7">
@@ -209,7 +232,7 @@ $(function(){
 </form>
 <?php }else{ ?>
 <h2 class="page-header">論壇</h2>
-<form class="form-inline" method="POST" action="forum.php?newblock">
+<form class="form-inline" method="POST" action="forum.php?newblock&<?php echo $_SESSION['Center_Auth']; ?>">
 	<div class="input-group">
 		<input class="form-control" name="blockname" type="text" placeholder="區塊名稱" required="required">
 		<span class="input-group-btn">
@@ -263,7 +286,7 @@ $(function(){
 				<td>
 					<a class="btn btn-info" href="forum.php?edit=<?php echo $_forum['row']['id']; ?>">編輯</a>
 					<?php if($_forum['num_rows']>1){ ?>
-					<a class="btn btn-danger" href="forum.php?delblock=<?php echo $_forum['row']['id']; ?>">刪除</a>
+					<a class="btn btn-danger" href="forum.php?delblock=<?php echo $_forum['row']['id'].'&'.$_SESSION['Center_Auth']; ?>">刪除</a>
 					<?php } ?>
 				</td>
 			</tr>
